requirements.txt
Django~=5.2.6
djangorestframework~=3.16.1
requests~=2.32.5
pillow~=11.3.0
supabase~=2.19.0
django-filter~=25.1
reportlab~=4.4.4
dj-config-url~=0.1.1
python-dotenv~=1.1.1
gunicorn~=21.2.0
django-cors-headers~=4.3.1
psycopg2-binary~=2.9.9
whitenoise~=6.6.0
# Solo dependencias del backend Django, sin librerías de IA ni FastAPI.
